chore: initialize Python client library for Authentik
This repository is based on the official ones for Golang and Swift and
the Makefile is mainly based on the goauthentik/client-swift one.

We made this repository because we need an Authentik Python library for
our internal tools, but we would like Authentik to take the ownership of
this repo and make it official. This is why we didn't make any release
nor pip publish for this repo.